+++ GdanskExplorer/Controllers/TripController.cs
@@ -1,5 +1,6 @@
 using System.Xml;
 using AutoMapper;
+using GdanskExplorer.Achievements;
 using GdanskExplorer.Data;
 using GdanskExplorer.Dtos;
 using GdanskExplorer.Topology;
@@ -21,15 +22,17 @@ public class TripController : ControllerBase
     private readonly UserManager<User> _userManager;
     private readonly GpxAreaExtractor _areaExtractor;
     private readonly IMapper _mapper;
+    private readonly AchievementManager _achievementMgr;
 
     public TripController(ILogger<TripController> logger, UserManager<User> userManager, GExplorerContext db,
-        GpxAreaExtractor areaExtractor, IMapper mapper)
+        GpxAreaExtractor areaExtractor, IMapper mapper, AchievementManager achievementMgr)
     {
         _logger = logger;
         _userManager = userManager;
         _db = db;
         _areaExtractor = areaExtractor;
         _mapper = mapper;
+        _achievementMgr = achievementMgr;
     }
 
     [HttpPost("new")]
@@ -107,13 +110,24 @@ public class TripController : ControllerBase
             );
 
             // purge all the previous ones for this user /shrug
-            await _db.DistrictAreaCacheEntries.Where(x => x.UserId == user.Id).ExecuteDeleteAsync();
+            if (await _db.DistrictAreaCacheEntries.AnyAsync(x => x.UserId == user.Id))
+            {
+                await _db.DistrictAreaCacheEntries.Where(x => x.UserId == user.Id).ExecuteDeleteAsync(); 
+            }
             // add new ones
             await _db.AddRangeAsync(newAreaCacheEntries);
             
 
             // save everything and return
             await _db.AddRangeAsync(dbTrips);
+            
+            // achievement check
+            await _db.Entry(user).Collection<Achievement>(x => x.Achievements).LoadAsync();
+            var remainingAchievements = _db.Achievements.Where(a => !user.Achievements.Contains(a)).ToList();
+
+            var newAchievements = _achievementMgr.CheckUser(remainingAchievements, user).ToList();
+            await _db.AddRangeAsync(newAchievements);
+            
             await _db.SaveChangesAsync();
             return Ok(_mapper.Map<IEnumerable<TripReturnDto>>(dbTrips));
         }
@@ -125,7 +139,7 @@ public class TripController : ControllerBase
     }
 
     [HttpGet("id/{guid:guid}")]
-    public ActionResult<DetailedTripReturnDto> GetById([FromRoute] Guid guid)
+    public async Task<ActionResult<DetailedTripReturnDto>> GetById([FromRoute] Guid guid)
     {
         var trip = _db.Trips.FirstOrDefault(x => x.Id.Equals(guid));
 
@@ -133,7 +147,16 @@ public class TripController : ControllerBase
         {
             return NotFound();
         }
+        
+        var user = await _db.Entry(trip).Reference(x => x.User).Query().SimplifyUser().FirstAsync();
+        _logger.LogDebug("got trip owner = {User}", user);
+        trip.User = user;
 
+        _db.Entry(trip).State = EntityState.Unchanged;
+        _db.Entry(user).State = EntityState.Unchanged;
+
+        await _db.SaveChangesAsync();
+        
         return Ok(_mapper.Map<DetailedTripReturnDto>(trip));
     }
 }
